gervis-backend/pyproject.toml
[project]
name = "gervis-backend"
version = "1.0.0"
description = "SpecTalk backend — FastAPI voice agent with Gemini Live"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "firebase-admin==6.6.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.10.2",
    "pydantic-settings==2.6.1",
    "python-dotenv==1.0.1",
    "httpx>=0.27.2",
    "cloud-sql-python-connector[asyncpg]==1.13.0",
    "psycopg2-binary>=2.9.11",
    "google-adk>=1.0.0",
    "deprecated>=1.3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["auth", "api", "db", "middleware", "agents", "tools", "services", "ws"]

[tool.uv]
package = false
